refactor(ui): extract graph construction into a pure, testable module
Move all non-component graph logic out of experiment-graph.tsx and into
ui/src/components/visualizations/experiment-graph/graph-builder.ts.

Extracted:
- nodeBox, NODE_DIMENSIONS, layoutGraph (dagre call)
- Index builders: buildChildMap, buildDirectionsByParent,
  buildDirectionsBySpawnExperiment, buildExperimentsByDirection,
  buildQuestionsByDirection, buildExperimentsByPrimaryQuestion,
  buildFindingsByExperiment
- Helpers: countStatuses, rootExperimentsForGroup, countDescendants,
  projectNodeId, bucketProjectId
- Subtree builders: addExperimentSubtree, addQuestionSubtree,
  addDirectionSubtree
- New top-level `buildExperimentGraph(input)` that takes a single
  `BuildGraphInput` and returns `{ nodes, edges, droppedOrphanEdges }`

The subtree builders previously threaded 19 positional parameters
through every recursive call; they now take a single `BuildContext`
object. No behavior change — same nodes, same edges, same layout.

The validation filter that drops orphan edges (added in the previous
commit) moves with the builder so every call path funnels through it.
The component keeps the warn-on-dropped-edges telemetry via a small
useEffect.

experiment-graph.tsx is now 913 LOC (was 1448). Node-type components
(ExperimentNode, ProjectNode, etc.) stay here — they move to their own
sub-files in the next commit.

Local verification: `npx tsc --noEmit` clean, `npm run build` clean,
`npx vitest run` 203/203 pass.
